chore(frontend): remove old calendar and use new one
